added processing lorun logic to it's own file and simplified main function
