--- src/comparison_data.rs
@@ -1,4 +1,0 @@
-mod ComparisonData{
-
-}
-
